tst_QLocale: replace complicated ifdefery with macro
This stops the shotgun surgery necessary when the condition changes,
by centralizing the condition and assigning a symbolic name to the
feature is represents. The more so as this list has evolved over the
lifetime of the still-active branches: First it was only ICU, then
Windows was added, and then Mac.

Task-number: QTBUG-2163
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I0fa604f9726a9fcf12b90655a5621b4169a7199b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>