QMacStyle: tweak NSTextFieldCell's drawing
This cell is drawing a frame/outline around a text field (a line edit in Qt).
Starting from macOS Tahoe this frame is slightly offset and with
larger size than requested (probably because now text field has
rounded corners and also the outline is thin and barely visible),
and as a result clipped by the current context's clip. In the
'Light' theme QLineEdit is completely invisible (unless it has a
focus). To address this issue we adjust the rect we draw line edit
with so it's the same as prior to Tahoe.

Pick-to: 6.10 6.9 6.8 6.5
Fixes: QTBUG-138465
Change-Id: I04e900ec4cc805d14a0774b5f3108d071c378789
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>