doc: Reword QGestureRecognizer docs slightly; fix a comment
"The application" was a bit vague: as a Qt user, I write an application;
but here, the QApplication object is what was apparently meant (although
by saying that, we omit the private detail that QApplicationPrivate has
a pointer to QGestureManager which actually takes ownership).
Investigation prompts changing the comment by the instance() method:
it's declared in QGestureManager, defined in qapplication.cpp.

Task-number: QTBUG-123585
Pick-to: 6.10 6.9 6.8
Change-Id: I50900697db8798c1ab4b9c5abaa9d2102a3cbcf4
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>