Properly guard supportedDragActions in QTreeWidget and QTableWidget
Guard Q_PROPERTY, the declarations and the definitions of the getters
and the setters with QT_CONFIG(draganddrop).

Both widgets use underlying models (QTreeModel and QTableModel) that
are based on QAbstractItemModel. The latter has a virtual method
supportedDragActions(), which is not guarded by any config.
As a result, this patch also has to provide fallback implementations
for the overrides of this method. The implementations simply return
Qt::IgnoreAction.

Amends 7d0017cda8fde28a1130feaeecf41010b40e3cb3 and
760579204e314cc4e0fdc7b8b8bd8d8706095a38.

Also provide fallback implementations for QTreeModel

Pick-to: 6.10
Change-Id: I9ff2c7112d1b3b7ad23347a03c5e29cf0d115935
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>