Warn qt6_add_resource(<target> ...) is called prior the target creation
qt6_add_resource function has two signatures that have completely
different processing. Since there is no explicit indication of what
signature should be used, qt6_add_resource checks if the first argument
is an exising target or not. This is not safe, since users may easily
mess the call order, so the target is not created when calling the
respective qt6_add_resource(<target> ...) function, and this leads to
unclear error. Since the arguments of two signatures differ
significantly, we may predict this situation and warn user, by checking
the use of the 'qt6_add_resource(<target> ...)' signature arguments in
the second 'qt6_add_resource(<outputfile> ...)' variant.

Pick-to: 6.10 6.9 6.8
Change-Id: I11df9ae56186cd56d9aac481a14b272f76716937
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>