CMake: Warn about public modules linking to private modules publicly
Public modules should not link publicly to private modules.
Doing that creates a public dependency on the private module CMake
packages. That forces user projects to automatically find the
private packages and make Private module targets and headers available
to the user project, without them explicitly opting in.

When building a developer-build Qt, show a warning when such a case is
detected, so that each repository can be fixed to not do this.

Add an opt out QT_WARN_PUBLIC_MODULE_LINKS_PRIVATE_MODULES_PUBLICLY
variable which can be set to OFF in the repo '.cmake.conf' file.

Pick-to: 6.10
Task-number: QTBUG-138712
Change-Id: I4cc5fba1a18f60c94b5e4c3f17432eba397029c6
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>