Remove documentation of non-existing QWeakPointer members
For historical reasons, qdoc sees a different declaration of
QWeakPointer than the compiler, and when we removed deprecated members
from the latter in e40320c552a9ccd4c1879a4da13c8d909b2cd21e for Qt 6.0,
we evidently forgot the former.

Remove those now from the declaration that qdoc sees as well, so that
we don't claim that they exist as deprecated members.

Pick-to: 6.5 6.8 6.9 6.10
Change-Id: I1167a9317df727054c0b0bc447083d64b59087bc
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>