Remove references to ConnectionMonitor from NetworkConnectionChannel
QNetworkConnectionMonitor is a private class that has been supported on
Windows and Apple. With the replacement of the Apple backend the Apple
support is removed, leaving only Windows to support it.

The plan is to remove QNetworkConnectionMonitor in a series of patches.
In this patch remove the references to QNetworkConnectionMonitor from
QHttpNetworkConnectionChannel.

Task-number: QTBUG-132137
Change-Id: Ibe8e4ea2816525b9c40acefc664d520fcbd9ff67
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>