Limit OpenGL software backend check to macOS Tahoe
Using > QOperatingSystemVersion::MacOSSequoia would include 15.0.1 and
up.

As of macOS 26 beta 3, the OS reports the major version as 26, even
when the app is built against a macOS 15 SDK, so we don't need to
account for the version being 16 in some cases.

Amends a9ca1aef2291abea36bba9b934f154ea9866337d

Change-Id: I9a2519ac20ffa0e10bdf258d1b12957f2b70c5dd
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>