QFileSystemEngine/BSD: check UF_HIDDEN on the symlink only
The property applies to the directory entry itself, which may be a
symlink. We don't really care if a symlink points to a hidden file, the
same way we wouldn't care if a symlink pointed to a file whose basename
started with a dot.

[ChangeLog][QtCore][QFileInfo] Fixed a bug that caused isHidden() to
return true on BSDs and Apple OSes for non-hidden symlinks that pointed
to hidden targets.

[ChangeLog][QtCore][QFileInfo] Fixed a bug that caused isHidden() to
return false on non-Apple BSD systems for symlinks that were hidden
using the chflags(1) tool with the -h flag.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I4a9a02621b577cded691fffd6968488cb047b1b3
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>