tst_QStringApiSymmetry: Pull collation locale from QLocale::system()
Instead of setlocale(LC_COLLATE, nullptr), which doesn't reflect the
system locale on all platforms, e.g. on Darwin.

This is important as we will end up in platform code later on when
doing string comparisons, and that code does reflect the system locale.

A mismatch here would end up failing localeAwareCompare_QString_QString
for C@auml.lt.aring as we would think we had a C locale, when the
system locale was actually en_NO.

Pick-to: 6.10 6.9 6.8 6.5
Done-with: Edward Welbourne <edward.welbourne@qt.io>
Change-Id: I24ed25b355772817d467c1585662d6abc25ff20c
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>