Revert "Fix flaky tst_QWidget::synthMouseDoubleClick on Opensuse"
This reverts commit d90dfe05609d48dba8767576a3d00d638d6fffc3.

Reason for revert: Fix flakiness on Opensuse, but create flakiness and
failing on Windows. The patch ensures that the window is positioned
in the top left corner of the screen. However, two tests from
tst_QWidget-taskQTBUG_27643_enterEvents and underMouse- which run
before synthMouseDoubleClick, place the cursor in the top left corner.
As a result, an unexpected mouseEvent is triggered when
synthMouseDoubleClick is executed, causing the test to fail due to
the altered cursor position.

Change-Id: Ic73ad7ca86f9ec42588b31ab067195e1d17e2bd5
Reviewed-by: Marc Mutz <marc.mutz@qt.io>