QLocale/Win: move Norwegian legacy codes into windows_to_iso_list
The id -> name lookup is using lower_bound(), so it will reliably find the first-listed of adjacent equivalent keys, which for both cases remains the new spelling. But by inserting the legacy spelling after the new one, the loop in qt_inIsoNametoLCID() will naturally find them. Yes, finding these two locale-ids will now be quite a bit slower, but everything else will be a bit faster. If this operation is speed-critical, we would have added an index (in the database sense) on iso_name, not just on windows_code.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Iee19bb664fa128d59dc5fc2a055c53d7da18b7c5
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>