QFuture: fix UB (uninit'ed member) in const_iterator default ctor
Coverity complained that the default ctor didn't initialize the
iterator's data members, and it's correct, of course.

What makes matters worse: by implementing the default constructor like
that, we break the C++ guarantee that value-initialized iterators must
compare equal.

Fix by = default'ing the constructor. That leaves default-construction
to establish only the partially-formed state, as before, but
value-initialization now creates a well-formed object, ready for
equality comparison.

Add a test.

Amends the start of the public history.

Coverity-Id: 87769
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I4d7af36937946dd58d52aaafae4dd48a60c00cf0
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>