QWidgetBaselineTest: Grab widget via render() instead of grab()
Going via QWidget::grab() will render to a QPixmap, which on most
platforms is going to be QRasterPlatformPixmap. The format of the
underlying QImage in that case is resolved from the primary screen,
which most platforms report as QImage::Format_RGB32.

This means we will not be able to detect the difference between
black and fully transparent areas in the rendered result, which
has been observed to be a problem for widget styles that wrongly
render their controls using source composition.

To solve this we render via QWidget::render() instead, to an image
with an alpha format.

Change-Id: I078f062b662590598c3597797abc2bfa78eaf86a
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>