QWidgetBaselineTest: Make test widget frameless and no drop shadow
And separate it from the background by making the background a
checkerboard pattern.

Avoids issues with rounded window frames contributing to the
results. Also, by making the window frameless we will not get
a default window background color on macOS, which means that
any issues with transparent rendering will be visible when
running the tests locally, as the checkerboard background
will shine though.

Change-Id: Icd501eeb149d2dbfb184c91e8eac3a531d09408d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>