Add EDOOM handling to VxWorks
VxWorks needs to have some extra handling for process shutdown when
threads are involved.
So when poll fails e.g. due to filehandles being corrupted or because of
running out of memory, these errors are ignored. Instead of ignoring the
errors the they should be reported and the eventloop should be stopped
if poll_exit_on_error is set.

Task-number: QTBUG-138831
Pick-to: 6.8 6.9 6.10
Change-Id: I1ebef63569d2ff4a18603d4984bba4abc4f033e8
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>