cmake: Persist macOS SDK name, to work around CMake 4.x
CMake 4 stopped passing explicit toolchain and SDK paths, which
results in lots of sharp corners.

For example, with xcode-select pointing to /Applications/Xcode.app
but with Xcode command line tools installed, CMake will build with
/usr/bin/c++, which xcrun resolves to the Xcode toolchain, but
as CMake doesn't pass a sysroot, it ends up using the command line
tools SDK. This is bad and surprising when the selected Xcode
version's SDKs doesn't match the command line tools.

CMake 4.0.2 was supposed to fix this, but it seemingly doesn't
work. As a workaround, set CMAKE_OSX_SYSROOT for macOS as well,
which triggers CMake or xcrun to pass a sysroot, giving us what
we want.

Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ia6d051477a48c4c8e5e90feae67cc8d2734d8e1b
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>