QLocale/Unix: enable -Wswitch in QSystemLocale::query()
The default: case label prevents the compiler from complaining about
unhandled enumerators in QueryType, of which there are six in this
case.

Add them explicitly and remove the default as a static assertion
(assuming -Werror=switch is enabled; which is the case at least for
-developer-build) that all enum values are handled.

Amends the start of the public history.

Found while fixing Coverity issue 455893.

Coverity-Id: 455893
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ia9711422cb5495f3322d8ad4701cbb4f35755474
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>