Mark QTranslator as security-critical
It parses .qm files, and even contains a byte code interpreter for
picking the correct plural form.

Even though .qm files are considered to be a trusted part of the
application, trust isn't a determining factor for security
criticality.

A user of QTranslator depends 100% on the component to do its job of
parsing data external to the process, with no error reporting other
than a bool. This makes the component security-critical.

The two header files only contain declarations, so are
security-significant (the default) with reason:header-decls-only.

Task-number: QTBUG-135190
Pick-to: 6.10 6.9 6.8
Change-Id: If6778fea42a60e4414624e7aaffb569583cebdd6
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>