Mark qvsnprintf as security-critical
This is borderline significant-only, because we only do minimal
processing here, and only on Android and WebAssembly (on other
platforms, we merely delegate to C's vsnprintf()).

But if you asked 100 programmers whether snprintf() is
security-critical, you would probably get 95 "yes" back, so mark it as
such. It's small, after all. Better safe than sorry.

There is not dedicated header for this TU (the functions are declared
in qbytearrayalgorithms.h, whose marking is for a separate patch).

Task-number: QTBUG-135195
Pick-to: 6.10 6.9 6.8
Change-Id: I45eaf0f240a57d92a71f83a7cd87b4d77dffe883
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>