QRM: reuse read/writeRole helpers in itemData/setItemData
There's no reason to duplicate the logic and specialization for QObject,
we do all of that in the helpers already. The helpers also fall-back to
Qt::DisplayRole when a property-lookup for Qt::EditRole fails, which we
also want for itemRole - if roleNames() includes that role, then it
should be part of the map of values returned.

This requires a small adjustment to the test, which already takes into
account that Qt::EditRole is faked in our implementation anyway.

Pick-to: 6.10
Change-Id: Id3d77cf7e83b4220ab2ac8aae7bba2a9de7d16d8
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>