HTTP: remap Custom operations to known methods
Some users are doing "CustomOperation" and then specifying GET, PUT, etc
To make it a bit easier to work with we should remap it to the known
methods. Especially since it makes an impact on caching and
redirections.

[ChangeLog][QtNetwork][QNetworkAccessManager] Using CustomOperation with
GET, PUT, POST, DELETE or HEAD will now be remapped to the corresponding
operation. This might affect caching and redirection handling.

Change-Id: I21b2ee950e73c8832a6c39accbb8c9acdae42a6e
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>