tst_qnetworkreply_local: shorten the local socket names
It seems to be failing with some test cases on some Linux systems.
It doesn't actually fail to listen on the server-side, but it seems the
client-side socket instead fails to connect.

Instead of using the full test case names and process id, hash it
together and use the HEX of that instead for the local socket
identifier.

Pick-to: 6.10 6.9 6.8
Change-Id: Ibcd0f6af29652d68470596efe84fba954cda57ce
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>