QLocale: add an easy missing std::move()
Coverity complained, and it's trivial to see, that a move is missing
in the assignment of local 'converted' to 'digits'. It's trivial
because the scope of 'converted' ends right afterwards.

Coverity complains about more missing moves in qlocale.cpp, but the
others are more complicated, so need a separate patch.

Amends ed2b110b6add650954dc102a0317c14ff826c677.

Coverity-Id: 427079
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: Ic124b2c9d9318cc7194d2f2affc8bd522be53fc7
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>