QDirListingPrivate: init() now takes no args
It had a bool arg mimicking the original QDirIterator code. This was
needed because QDI had a ctor that took a QDir (and QDI itself at one
point in the distant past inherited QDir), and the resovleEngine
arg was set to true/false depending on the QDirPrivate::fileEngine
pointer. This isn't the case any more, and looking back,
QDirListingPrivate::init() has always been called with a true arg.

Change-Id: Ib488ca459234e22653c6efe4013adb1343a2d19c
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>