Fix crash when sending ObjectDestroyed events without an a11y interface
The a11y test library copies events, there is a case where we send
ObjectDestroyed on an object, as opposed to the expected a11y
interface.

When this happens, the program will crash on a null pointer
read.

The fix is to support these two scenarios:
1) (object == nullptr, m_uniqueId set)
2) (object != nullptr, m_childId set)

For the first case, we no longer have access to the object,
so event->object() will be nullptr. This is a change, and happens
when the event is sent from the various destructors.

For the second case, which happens when the object is still
alive, we set m_object, and QAccessibleEvent::accessibleInterface
will do the correct thing and return the interface, this is also a
change, earlier we would try to access m_uniqueId which was
aliased by childId(-1).

Fixes: QTBUG-138824
Pick-to: 6.10
Change-Id: If8e4f8d0e26c3c5793f876b1c25b0653621af1e9
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>