Android: fix crash with -no-accessibility and a11y enabled on-device
When Qt is built with -no-accessibility and the device has
accessibility enabled (e.g. Talkback), it will lead to a crash
from calling an unregistered native method. Now, this adds a small
native method that's registered on library load that returns whether
Qt is configured with accessibility or not which decides whether
to init the accessibility layer from Java or not.

Pick-to: 6.10
Fixes: QTBUG-138894
Change-Id: I6d5f86b8777c12d3b82672a221827dffcfccbbf2
Reviewed-by: Petri Virkkunen <petri.virkkunen@qt.io>