Remove spurious '\0' termination from a CharBuff
The CharBuff, a QVarLengthArray<char, 256>, knows how big it is and is
used by callers who use its start and size, rather than relying on
'\0'-termination (although they did if you go far enoubh back in
history). Two callers convert it to QByteArrayView, which took care of
ignoring the dangling '\0' (despite it being within size, rather than
just after it). The third had to subtract one from the size to correct
for the '\0': it now no longer needs to do that.

Pick-to: 6.10
Change-Id: I09870d3f4f6871f37106d82f251611db45eb2101
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>