iOS: Add universal link cold-launch handling
Starting from Qt 6.8 the iOS support is based on
UIScenes, but it seems the cold-launch case for
handling universal links was missing. The application
would launch but the URL data is not caught, making
the applications unable to receive the universal link
URL.

By experimentation there are four ways for the iOS
window scene delegate to receive URLs:

1) Custom-scheme URI when application is running:
   openURLContexts() called by iOS
2) Universal link (https) when application is running:
   continueUserActivity() called by iOS
3) Custom-scheme URI cold-launch:
   willConnectToSession() called by iOS. The URL
   is provided in connectionOptions.URLContexts
4) Universal link (https) cold-launch:
   willConnectToSession() called by iOS. The URL
   is provided in connectionOptions.userActivities

The handling for the last one was missing and is
added by this commit.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-138488
Change-Id: I10ea0e3f98d688a83e6fa02d7bb78497f48908ff
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>