Android: 16KB page size support for building Qt for Android
This patch introduces android_16kb_pages feature.

Feature is needed for Qt versions supporting Android 15 or newer
which use NDK 25 or later.

Feature adds linker options to build shared libraries (.so)
with 16KB page size instead of 4KB page size. The 16KB page size
is compatible with 4KB page size.

Adding the linker flag is applicable only for 64-bit target
architectures: x86_64 and arm64-v8a when using NDK 25, 26 or 27.
NDK 28 and later will have the feature built-in so the linker flags
and note are not needed to be used with those versions.

It is applied to both Qt build and Qt Application builds.

Task-number: QTBUG-131514
Task-number: QTBUG-134093
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I8af9c3b392a661995c2eb46b568a40b94ee62aaa
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>