moveToTrash/XDG: allow $XDG_DATA_HOME/Trash to be a symlink
This is in the user's $HOME dir and thus entirely controlled by the
user, so we don't need to apply the strict protections that the XDG spec
requires of other dirs. The trash spec[1] being one long section without
structures made the reading ambiguous whether the checks should also be
applied to $HOME.

This commit also allows the "info" and "files" subdirs to be symlinks,
but I'm not testing that.

[ChangeLog][QtCore][QFile] Fixed a bug that caused moveToTrash() to
disallow trashing to trash bins using the XDG Trash Specification when
the trash path was a symlink.

[1] https://specifications.freedesktop.org/trash-spec/1.0/

Pick-to: 6.8 6.9 6.10
Fixes: QTBUG-138984
Change-Id: I4f05391082bc607389e0fffd06a9b72fdd098f22
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: David Faure <david.faure@kdab.com>