QCOMPARE: print the sign for negative zeroes and NaN
To do that, we cannot convert to double from float or std::float16,
because the conversion can normalize the output to standard zero or
quiet NaN.

We also need to use signbit() because the comparison is always false
otherwise: -0 is equal to 0 (not less than) and all comparisons on NaN
produce false. Moreover, like fpclassify(), signbit() is guaranteed by
the C standard not to raise an NaN consumption exception.

ISO C23 F.3 paragraph 6:
> The C classification macros fpclassify, iscanonical, isfinite, isinf,
> isnan, isnormal, issignaling, issubnormal, iszero, and signbit provide
> the ISO/IEC 60559 operations indicated in Table F.2 provided their
> arguments are in the format of their semantic type. Then these
> macros raise no floating-point exceptions, even if an argument is a
> signaling NaN.

Change-Id: I384599c59c440f36836afffde93b27fdee7a01a6
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>