Windows with D3D: Guard vsync update deliveries more
Flip m_vsyncUpdatePending back to "ready" only on the
gui thread, when the queued metacall is processed.
Also have a third state, "posted", to indicate that
the invokeMethod call has been made but not yet
processed on the gui thread.

Otherwise certain Quick applications can end up in
a situation where there are an ever growing number
of metacalls queued, not having the chance to
process them all if new requestUpdate() calls are
made continuously as a reaction to Qt Quick doing
a new frame.

Done-with: Christian Strømme <christian.stromme@qt.io>
Fixes: QTBUG-139014
Pick-to: 6.10 6.9
Change-Id: I1ae4f28ee6a52d66cc4349bf930c5e873a7e299c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>