Add macOS backend for QRandomAccessAsyncFile
The backend uses Grand Central Dispatch to perform async IO.

One complication is that the GCD APIs do not allow to cancel a
signle IO read/write operation, instead it's only possilbe to
close the related IO channel. Luckily, it's possible to have
multiple IO channels referencing the same file descriptor.

As a result, we end up duplicating IO channels for each async
operation, which makes the code a bit more complicated.

The implementation uses QWaitCondition to make sure that the
async operation is completed or canceled before removing the
relevant QIOOperation or QRandomAcessAsyncFile instance.

Another tricky part was the implementation of flush(). The
docs for dispatch_io_barrier claim that it applies to a file
descriptor, not a specific channel, and thus acts as a barrier
across all channels that share the same file descriptor.
However, in practice it does not work! As a result, the patch
implements an operation queue, and schedules the barrier
operations only when all previous operations are completed.
Similarly, no new operations are scheduled until the barrier
operation is completed.

The same logic is applied to the async open() operations that
are also considered to be barrier operations.

Fixes: QTBUG-139005
Change-Id: Iac448d9460c64dd18dc3b15fc13d5145895a3c3b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>