Doc: Document QTouchEventWidgetSequence, QTestResult
In Qt 6, widget-specific functions for creating touch event sequences
were moved to a new class, QTouchEventWidgetSequence. We mention this
class in our porting guide and document its member functions, but
the class itself was not documented.

Adding the \class topic clears QDoc warnings such as:

(qdoc) warning: No output generated for function
    'QTest::QTouchEventWidgetSequence::press()' because
    'QTouchEventWidgetSequence' is undocumented

Similar warnings are generated for class QTestResult with a private
implementation; add \class topic for it but mark it \internal.

Pick-to: 6.10 6.9 6.8
Change-Id: I642dc4b160601247a6938b228caff788f655c866
Reviewed-by: David Boddie <david.boddie@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>