CMake: fix the reduce_relocations test
This test was apparently never properly ported from qmake, because it
forgot to check that the linker supports the flag that it meant to use.
Fix it by checking for both flags that we need to support the feature,
-Bsymbolic-functions and --dynamic-list.

This also fixes the incorrect GCC checks, since this is not a compiler
feature, but a linker one.

Pick-to: 6.10
Change-Id: I92f6bb1e213e335b5c94fffd3ea509c53b847ee6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>