QRM: mark structure and size functions as final
QRM needs to be able to rely on row/columnCount, functions implementing
parent/child navigation, the construction of the QModelIndex, and APIs
for inserting, removing, or moving rows and columns using the underlying
data structure in a consistent way.

A subclass trying to come up with a different logic, or modifying
results, will throw the rest of the implementation easily off to the
point where we end up operating with invalid iterators or bad pointers.

So, mark those APIs that shouldn't be overridden further as final. We
can relax that restriction in the future if it's really needed.

Pick-to: 6.10
Change-Id: I5e2195a631ae148cb49e8d1a190378e415396a5c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>