Destroy the QUnifiedTimer when the Q*Application gets destroyed
Cleanup the thread local instance of the QUnifiedTimer when
the QCoreApplication::aboutToQuit signal gets emitted.
This should make sure the QUnifiedTimer for the main thread
is destroyed and doesn't try to run any AnimationJobs which are
already deleted.

The quit() may be invoked from an animation. In that case, the
QUnifiedTimer still needs to survive its currently running
updateAnimationTimers(). Therefore, use deleteLater() rather than
deleting directly.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-136629
Change-Id: I1608e566ee9b3ab313e82d3233fe455fd488b166
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>