Fix '\0'-termination issue in qTzName
Passing const char * (with no size) to QString::fromLocal8Bit() forces
use of the QByteArrayView overload of the latter, which relies on '\0'
termination. So using std::min(strlen() + 1, sizeof()) as the size to
memcpy() may copy a truncated version of the string (already bad) and,
failing to '\0'-terminate, trigger UB in QBAV's constructor (much worse).

Make strlen() < sizeof() a precondition of copying and join the other
failure paths that return a null QString in the (admittedly unlikely)
situation where it fails. In the process, unify with the MS
code-branch since both now know the size copied on success, so can
pass it to fromLocal8Big(const char *, qsizetype) and avoid the
problem. The code dates back to 6.6, so only pick back to 6.8.

Pick-to: 6.10 6.9 6.8
Change-Id: If7ceb4ce600a8d383e08c4b798fa720f33f4c9ff
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>