Update byte-ordered QPixelFormats to be QPixelFormat::BigEndian
If a byte-ordered format is read in larger chunks, such as when passing
a 32-bit format to platform APIs as a single integer, the interpretation
of the bytes needs to be in big endian.

Reflect this though the built in pixel formats for QImage formats,
as well as the default behavior of the QPixelFormat constructors
and factory functions.

Also clarify this in the documentation.

[ChangeLog][QtGui] QPixelFormats with typeInterpretation()
QPixelFormat::UnsignedByte now reflect a byteOrder of
QPixelFormat::BigEndian, to highlight the fact that these
formats need to be read as such if not read in a byte-by-byte
manner.

Change-Id: I03baee3a50e7c7d94bc309933e00566653d42de6
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>