tests: buffer overflow warning in wayland reconnect test
The display_name buffer size was insufficient for the maximum possible
output of snprintf with format "wayland-%d". The compiler warned that
output could be 10-20 bytes but the destination buffer was only 16 bytes.

Increase buffer size from 16 to 32 bytes to accommodate the worst case
scenario (e.g., "wayland--2147483647") and provide safety margin.

Change-Id: Id980a0d897f06d1a0c1f2cb2afba9f6cd36c2988
Reviewed-by: David Edmundson <davidedmundson@kde.org>