Improve wording in docs of QTimeZone::TimeType
This type is only used to configure displayName(), so express it in
terms that'll make sense to those reading its doc. I'm not sure what
the prior opening wording was even saying. Introduce by explaining
what it's about, thereby establishing a context in which it's possible
to say what distinguishes the enum expresses. Prune extraneous
verbiage from examples.

Pick-to: 6.10 6.9 6.8
Change-Id: I4adeb954d20b5e48e3a09142cd24cb180d82a444
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>