qmake: Make ccache feature available on windows hosts
Previously the ccache feature was only available on Unix-like
hosts. This change makes it available on Windows hosts as well.

Guard the ccache sloppiness assignments not to run on Windows hosts,
due to syntax errors in the Windows shell.

Also make sure to only wrap the compilers when targeting msvc,
otherwise sccache specifically will complain when executing link.exe,
which is a linker. ccache doesn't have this issue.

Pick-to: 6.8 6.9 6.10
Task-number: QTBUG-139139
Change-Id: Icf314cf53197d6372c2794ecfb52786c1f9ee6d4
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>