QBuffer: replace Q_PRIVATE_SLOT with QMetaObject::invokeMethod
There's a single call site, so this moves the effect closer to the cause
and avoids a possible mistake in the string name. This delayed emission
appears to exist so we "compress" the emission of those two signals and
emit them only once, however many write() calls there may have been.

We lose the vtable dispatch with this change, but that would only be a
problem if a parent destructor (i.e., ~QIODevice()) posted events before
~QObject() cleared them. That is not the case now and unlikely to become
so.

Change-Id: I33cac248a651c88c7eb5fffd259f6b58a56d5d8a
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>